Task: Organize S3 Classes in Amazon Provider
