+++ tests/providers/amazon/aws/operators/test_s3_list_prefixes.py
@@ -19,7 +19,7 @@
 import unittest
 from unittest import mock
 
-from airflow.providers.amazon.aws.operators.s3 import S3ListPrefixesOperator
+from airflow.providers.amazon.aws.operators.s3_list_prefixes import S3ListPrefixesOperator
 
 TASK_ID = 'test-s3-list-prefixes-operator'
 BUCKET = 'test-bucket'
